\name{directPA-package}
\alias{directPA-package}
\docType{package}
\title{
Direction Pathway Analysis Package
}
\description{
The directPA-package is designed for analysing pathways in experiments with multiple perturbations. The combination effects of
different treatments are tested by rotating polar coordinates in two-dimentional space when the experiment contains two perturbations
and corresponding controls, or spherical coordinates in three-dimensional space when the experiment contains three perturbations and 
corresponding controls.
}
\details{
\tabular{ll}{
Package: \tab directPA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-21\cr
License: \tab GPL-3\cr
}

}
\author{
Pengyi Yang <yangpy7@gmail.com> & Ellis Patrick <epatrick87@gmail.com>
}
\references{
Pengyi Yang, Ellis Patrick, Shi-Xiong Tan, Daniel J. Fazakerley, James Burchfield, Christopher Gribben, Matthew J. Prior, 
David E. James, Yee Hwa Yang, Direction pathway analysis of large-scale proteomics data reveals novel features of the insulin action pathway, submitted.
}
