% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geneStats.R
\name{geneStats}
\alias{geneStats}
\title{Gene Level Statistics}
\usage{
geneStats(T, method = "OSP")
}
\arguments{
\item{T}{a vector of statistics (z-scores) with each element correspond to a treatment vs control comparison.}

\item{method}{the p-value integration method for combining accross multiple treatments. Available methods are
Stouffer, OSP, Fisher, and maxP. The default method is OSP.}
}
\value{
a p-value after integration across treatments.
}
\description{
Takes a vector of statistics with each element corresponds to a treatment vs control comparison,
and calculates a combined statistics accross multiple treatments.
}
\examples{
# Load the example data.
data(PM)

# (1) For three perturbatins vesus controls, use three dimentional rotation.
Tc <- cbind(Ins, Wmn, MK)
}

