\name{HEK}
\docType{data}
\alias{HEK}
\title{Phosphoproteomics of HEK-293E}
\description{
The data object contains all quantified phosphoryaltion sites in HEK-293E cells with Rapamycin or Torin1 treatment vs insulin stimulation. See
Hsu et al. Science, 332(6035), 1317-1322, 2011, for more details.
}
