% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbPlot3d.R
\name{perturbPlot3d}
\alias{perturbPlot3d}
\title{Perturbation Plot 3D}
\usage{
perturbPlot3d(Tc, annotation, minSize=5, ...)
}
\arguments{
\item{Tc}{a numeric matrix. The columns are genes or phosphorylation sites and the columns are treatments 
vs control statistics.}

\item{annotation}{a list with names correspond to pathways or kinases and elements correspond to genes or
substrates belong to each pathway or kinase, respectively.}

\item{minSize}{the size of annotation groups to be considered for calculating enrichment. Groups 
that are smaller than the minSize will be removed from the analysis.}

\item{...}{parameters for controling the plot.}
}
\value{
a list of coordinates for pathways or kinases
}
\description{
This function takes in a matrix of test statistics with two columns (3-dimensional space) and the 
annotation list such as pathway annotation or kinase-substrate annotation, and visualize the enrichment
of pathways or kinases in direction specific manner.
}
