\name{PhosphoELM.mouse}
\docType{data}
\alias{PhosphoELM.mouse}
\title{PhosphoELM annotations for mouse}
\description{
The data object contains the annotations of kinases and their conrresponding substrates as phosphorylation sites in mouse. It is extracted
from the PhosphoELM database. For details of PhosphoELM, please refer to the article: Phospho.ELM: a database of phosphorylation sites--update 2008.
Diella, F., Gould, C.M., Chica, C., Via, A. & Gibson, T.J. Nucleic Acids Res. 2008 Jan;36(Database issue):D240-4.
}
