\name{PM}
\docType{data}
\alias{PM}
\title{Plasma Membrame Protoemics Data}
\description{
The data object contains all quantified plasma membrame proteins with insulin simulation with and without prior treatments of MK or Wmn vs basal 
condition in mouse 3T3L1 cell lines. Please refer to the article: Yang et al. Bioinformatics, 30(6):808-14, 2014.
}
