\name{reactome.list}
\docType{data}
\alias{reactome.list}
\title{Reactome Database}
\description{
Reactome database in list format. Each list element contains an array of gene names corresponding to the genes contains in a pathway. The name of the list element is the pathway name.
}
\references{
Joshi-Tope, G., et al. Reactome: a knowledgebase of biological pathways. Nucleic Acids Research 33.suppl 1 (2005): D428-D432.
}
